Refactor Polish researcher
