+++ packages/yoastseo/src/languageProcessing/languages/pl/config/internal/auxiliaries.js
