+++ packages/yoastseo/src/languageProcessing/languages/pl/config/internal/participles.js
