+++ packages/yoastseo/src/languageProcessing/languages/pl/helpers/getSentenceParts.js
@@ -1,7 +1,7 @@
 import splitSentence from "../../../helpers/passiveVoice/periphrastic/getSentencePartsSplitOnStopwords";
 import arrayToRegex from "../../../helpers/regex/createRegexFromArray";
-import SentencePart from "../config/passiveVoice/SentencePart";
-import auxiliariesFactory from "../config/passiveVoice/auxiliaries.js";
+import SentencePart from "../values/SentencePart";
+import auxiliariesFactory from "../config/internal/auxiliaries.js";
 import stopwordsFactory from "../config/stopwords.js";
 
 const options = {
@@ -17,9 +17,9 @@ const options = {
  * Gets the sentence parts from a sentence by determining sentence breakers.
  *
  * @param {string} sentence The sentence to split up in sentence parts.
- * @param {Object} options The language options for which to get the sentence parts.
+ *
  * @returns {Array} The array with all parts of a sentence that have an auxiliary.
  */
-const getSentenceParts = function( sentence) {
+export default function getSentenceParts( sentence ) {
 	return splitSentence( sentence, options );
-};
+}
