+++ packages/yoastseo/src/languageProcessing/languages/pl/helpers/internal/stem.js
