+++ packages/yoastseo/src/languageProcessing/languages/pl/values/PolishParticiple.js
@@ -1,8 +1,9 @@
-import Participle from "../../../../../values/Participle.js";
-import checkException from "../../../../helpers/passiveVoice/periphrastic/checkException.js";
-import nonDirectPrecedenceException from "../../../../helpers/passiveVoice/periphrastic/freeAuxiliaryParticipleOrder/nonDirectParticiplePrecedenceException";
-import directPrecedenceException from "../../../../helpers/passiveVoice/directPrecedenceException";
-import { getFunctionWords } from "../functionWords.js";
+import Participle from "../../../../values/Participle.js";
+import checkException from "../../../helpers/passiveVoice/periphrastic/checkException.js";
+import nonDirectPrecedenceException from
+	"../../../helpers/passiveVoice/periphrastic/freeAuxiliaryParticipleOrder/nonDirectParticiplePrecedenceException";
+import directPrecedenceException from "../../../helpers/passiveVoice/directPrecedenceException";
+import getFunctionWords from "../config/functionWords.js";
 const {
 	cannotDirectlyPrecedePassiveParticiple: cannotDirectlyPrecedePassiveParticipleList,
 } = getFunctionWords();
